moved regressors to separated library
