+++ core/CMakeLists.txt
@@ -4,28 +4,18 @@
 
 project(iDynTree_Core CXX)
 
-set( PROJ_SOURCE_DIR ${kdl_codyco_SOURCE_DIR} )
-set( PROJ_BINARY_DIR ${kdl_codyco_BINARY_DIR} )
-
-option(KDL_CODYCO_ENABLE_RPATH "Enable RPATH for the library" TRUE)
-mark_as_advanced(KDL_CODYCO_ENABLE_RPATH)
-
 if(MSVC)
     set(CMAKE_DEBUG_POSTFIX "d")
 endif(MSVC)
 
-include_directories(include ${orocos_kdl_INCLUDE_DIRS} ${EIGEN3_INCLUDE_DIR})
+include_directories(include ${orocos_kdl_INCLUDE_DIRS}
+                            ${EIGEN3_INCLUDE_DIR})
 
-include_directories(include
-                    ./include/kdl_codyco/regressors ./include/kdl_codyco/regressors/dataset ${orocos_kdl_INCLUDE_DIRS})
-
-file(GLOB  KDL_CODYCO_HPPS
+file(GLOB  IDYNTREE_CORE_HDRS
      "include/kdl_codyco/*.hpp"
  )
 
-set(IDYNTREE_CORE_HDRS include/kdl_codyco/treeserialization.hpp)
-
-message("KDL_CODYCO_HPPS: ${KDL_CODYCO_HPPS}")
+#set(IDYNTREE_CORE_HDRS include/kdl_codyco/treeserialization.hpp)
 
 file (GLOB KDL_CODYCO_PRIVATE_INCLUDES
     "include/kdl_codyco/config.h"
@@ -33,10 +23,6 @@ file (GLOB KDL_CODYCO_PRIVATE_INCLUDES
 
 file(GLOB KDL_CODYCO_SRC
     "src/*.cpp"
-)
-
-file(GLOB DIRL_SRC
-    "src/regressors/*.cpp" "src/regressors/dataset/*.cpp"
 )
 
 SOURCE_GROUP("Source Files" FILES ${KDL_CODYCO_SRC} ${DIRL_SRC})
@@ -48,18 +34,6 @@ get_property(IDYNTREE_TREE_INCLUDE_DIRS GLOBAL PROPERTY IDYNTREE_TREE_INCLUDE_DI
 
 include_directories(${IDYNTREE_TREE_INCLUDE_DIRS})
 
-if(YCM_FOUND)
-include(YCMDefaultDirs)
-
-include(AddInstallRPATHSupport)
-add_install_rpath_support(BIN_DIRS "${CMAKE_INSTALL_PREFIX}/bin"
-                          LIB_DIRS "${CMAKE_INSTALL_PREFIX}/lib"
-                          DEPENDS KDL_CODYCO_ENABLE_RPATH
-                          USE_LINK_PATH)
-
-endif()
-
-
 add_library(idyntree-core ${KDL_CODYCO_SRC} ${DIRL_SRC} ${KDL_CODYCO_HPPS} ${KDL_CODYCO_PRIVATE_INCLUDES})
 
 target_include_directories(idyntree-core PUBLIC "$<BUILD_INTERFACE:${CMAKE_SOURCE_DIR}/include>"
@@ -67,24 +41,7 @@ target_include_directories(idyntree-core PUBLIC "$<BUILD_INTERFACE:${CMAKE_SOURC
 
 target_link_libraries(idyntree-core ${orocos_kdl_LIBRARIES})
 
-set_target_properties(idyntree-core PROPERTIES
-  SOVERSION "${KDL_CODYCO_VERSION_MAJOR}.${KDL_CODYCO_VERSION_MINOR}"
-  VERSION "${KDL_CODYCO_VERSION}"
-  COMPILE_FLAGS "${CMAKE_CXX_FLAGS_ADD}")
-
-
 set_property(TARGET idyntree-core PROPERTY PUBLIC_HEADER ${IDYNTREE_CORE_HDRS})
-
-if(${CMAKE_MINIMUM_REQUIRED_VERSION} VERSION_GREATER "2.8.12")
-    message(AUTHOR_WARNING "CMAKE_MINIMUM_REQUIRED_VERSION is now ${CMAKE_MINIMUM_REQUIRED_VERSION}. This check can be removed.")
-endif()
-if(CMAKE_VERSION VERSION_LESS 2.8.12)
-    if(NOT KDL_CODYCO_DISABLE_RPATH)
-        set_target_properties(idyntree-core PROPERTIES
-            INSTALL_NAME_DIR "${CMAKE_INSTALL_PREFIX}/lib"
-        )
-    endif(NOT KDL_CODYCO_DISABLE_RPATH)
-endif(CMAKE_VERSION VERSION_LESS 2.8.12)
 
 install(TARGETS idyntree-core
         EXPORT iDynTree
@@ -96,9 +53,3 @@ install(TARGETS idyntree-core
         PRIVATE_HEADER DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/kdl_codyco/impl)
 
 set_property(GLOBAL APPEND PROPERTY ${VARS_PREFIX}_TARGETS idyntree-core)
-
-
-if(IDYNTREE_COMPILE_TESTS)
-    add_subdirectory(tests)
-endif(IDYNTREE_COMPILE_TESTS)
-
