+++ regressors/src/DynamicDatasetFile.cpp
@@ -7,7 +7,7 @@
 #include <iostream>
 #include <sstream>
 
-#include <kdl_codyco/regressors/dataset/DynamicDatasetFile.hpp>
+#include <kdl_codyco/regressors/DynamicDatasetFile.hpp>
 
 namespace KDL {
 namespace CoDyCo {
