+++ regressors/src/dynamicRegressorGenerator.cpp
